requirements.txt
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
SQLAlchemy==0.9.8
alembic==0.7.4
kombu==3.0.33
psycopg2==2.6.1
